Fix log generation
We were generating a log based on all branches. This was fine for the
Linux kernel, but for MariaDB there are a bunch of branches that can
skew statistics, these fall into two categories:

1. Branches of feature development that are not in the main branches
2. Preview release branches

The latter can have duplicate commits with different histories which
cause duplicate commits in statistics.

We now filter by main development branches only.