Add categories support
The employers output now shows an extra column for category and there
is a new gitdm option `-A` which outputs a categories list in a separate
file. The employer to category map is defined as a `CategoryMap` file.
This needs to be first in the `gitdm.config` file, otherwise the data is
parsed without it.